DB call for insert user and check user

+++ app/src/main/java/model/UserCalorieCount.java
@@ -25,33 +25,33 @@ public class UserCalorieCount {
         this.date = date;
     }
 
-    public float getTotal_cal() {
+    public double getTotal_cal() {
         return total_cal;
     }
 
-    public void setTotal_cal(float total_cal) {
+    public void setTotal_cal(double total_cal) {
         this.total_cal = total_cal;
     }
 
-    public float getTotal_protien() {
+    public double getTotal_protien() {
         return total_protien;
     }
 
-    public void setTotal_protien(float total_protien) {
+    public void setTotal_protien(double total_protien) {
         this.total_protien = total_protien;
     }
 
-    public float getTotal_fiber() {
+    public double getTotal_fiber() {
         return total_fiber;
     }
 
-    public void setTotal_fiber(float total_fiber) {
+    public void setTotal_fiber(double total_fiber) {
         this.total_fiber = total_fiber;
     }
 
     private Date date;
-    private float total_cal;
-    private float total_protien;
-    private float total_fiber;
+    private double total_cal;
+    private double total_protien;
+    private double total_fiber;
 
 }
